LC-sqrtx [Runtime: N/A; Memory: N/A]

+++ binary-search/sqrtx.cpp
@@ -9,7 +9,7 @@ public:
         int ans = 1;
         while(low <= high)
         {
-            int mid = low + (high - low)/2;
+            long long mid = low + (high - low)/2;
             long long sq = mid*mid;
             if(sq == x)
             {
